Add support for gathering boost

+++ contracts/globals/players.sol
@@ -36,7 +36,7 @@ struct Attire {
   uint16 feet;
   uint16 ring;
   uint16 reserved1;
-  uint128 queueId; // Just here for packing purposes
+  uint80 queueId; // Just here for packing purposes
 }
 
 struct CombatStats {
